+++ human_sim/config/dev.exs
@@ -1,0 +1,13 @@
+# Human Simulator – dev config
+
+import Config
+
+config :human_sim, HumanSimWeb.Endpoint,
+  http: [ip: {127, 0, 0, 1}, port: 4000],
+  check_origin: false,
+  code_reloader: true,
+  debug_errors: true,
+  secret_key_base: "dev_secret_key_base_64_chars_minimum_for_phoenix_crypto",
+  watchers: [
+    esbuild: {Esbuild, :install_and_run, [:default, ~w(--sourcemap=inline --watch)]}
+  ]