+++ human_sim/lib/human_sim_web/components/error_html.ex
@@ -1,0 +1,11 @@
+defmodule HumanSimWeb.ErrorHTML do
+  use HumanSimWeb, :html
+
+  def render("404.html", _assigns) do
+    "Page not found"
+  end
+
+  def render("500.html", _assigns) do
+    "Internal server error"
+  end
+end