+++ human_sim/lib/human_sim_web/components/layouts.ex
@@ -1,0 +1,5 @@
+defmodule HumanSimWeb.Layouts do
+  use HumanSimWeb, :html
+
+  embed_templates "layouts/*"
+end